Rename MCDownloader to MojangApi and Skin to PlayerSkin.
