+++ chunky/src/java/se/llbit/chunky/entity/HeadEntity.java
@@ -19,20 +19,16 @@
 package se.llbit.chunky.entity;
 
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.net.URL;
-import java.nio.channels.Channels;
-import java.nio.channels.ReadableByteChannel;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.Map;
 import java.util.WeakHashMap;
-import se.llbit.chunky.PersistentSettings;
+import se.llbit.chunky.renderer.scene.PlayerModel;
 import se.llbit.chunky.resources.PlayerTexture;
 import se.llbit.chunky.resources.Texture;
-import se.llbit.chunky.resources.texturepack.EntityTextureLoader;
+import se.llbit.chunky.resources.texturepack.PlayerTextureLoader;
 import se.llbit.chunky.resources.texturepack.TextureFormatError;
 import se.llbit.json.JsonObject;
 import se.llbit.json.JsonValue;
@@ -42,8 +38,7 @@ import se.llbit.math.Transform;
 import se.llbit.math.Vector3;
 import se.llbit.math.primitive.Box;
 import se.llbit.math.primitive.Primitive;
-import se.llbit.util.MCDownloader;
-import se.llbit.util.Util;
+import se.llbit.util.mojangapi.MojangApi;
 
 /**
  * A mob head (skull) entity.
@@ -154,9 +149,9 @@ public class HeadEntity extends Entity {
   public static PlayerTexture downloadTexture(String skin) {
     return textureCache.computeIfAbsent(skin, (skinUrl) -> {
       PlayerTexture texture = new PlayerTexture();
-      EntityTextureLoader loader = new EntityTextureLoader(skinUrl, texture);
+      PlayerTextureLoader loader = new PlayerTextureLoader(skinUrl, texture, PlayerModel.STEVE);
       try {
-        File cacheFile = MCDownloader.downloadSkin(skin);
+        File cacheFile = MojangApi.downloadSkin(skin);
         try {
           loader.load(cacheFile);
           return texture;
